pyproject.toml
[project]
name = "erc-20"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "mamushi>=0.1.0",
    "pytest>=9.0.2",
    "ruff>=0.14.14",
    "titanoboa>=0.2.8",
]
